requirements.txt
# Sentinel-RAG: Military-Domain Adaptive GraphRAG

# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0

# ML / Embeddings
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0

# PDF Parsing
PyMuPDF>=1.23.0

# Vector Store
faiss-cpu>=1.7.0

# Graph
networkx>=3.0

# LangChain (for vanilla RAG baseline)
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langchain-openai>=0.1.0
chromadb>=0.4.0

# LLM Providers
openai>=1.0.0
google-generativeai>=0.5.0

# Evaluation (official RAGAS)
ragas>=0.1.0
rouge-score>=0.1.2

# Statistical Analysis
scipy>=1.10.0
statsmodels>=0.14.0

# Reporting
tabulate>=0.9.0
pandas>=2.0.0
